# api/openapi.yaml
openapi: 3.0.0
info:
  title: test-api
  version: "1.0"
  description: "Test api project"
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/tasks:
    post:
      tags: [tasks]
      operationId: createTask
      description: Создать задачу
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [team_id, title]
              properties:
                team_id: { type: integer, format: int64 }
                title: { type: string }
                description: { type: string }
                assignee_id:
                  type: integer
                  format: int64
                  nullable: true
      responses:
        "201":
          description: Задача создана
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Task" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    get:
      tags: [tasks]
      operationId: listTasks
      description: Список задач команды
      parameters:
        - name: team_id
          in: query
          required: true
          schema: { type: integer, format: int64 }
        - name: status
          in: query
          schema: { $ref: "#/components/schemas/TaskStatus" }
        - name: assignee_id
          in: query
          schema: { type: integer, format: int64 }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
        - name: offset
          in: query
          schema: { type: integer, minimum: 0, default: 0 }
      responses:
        "200":
          description: Страница задач
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TaskPage" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/tasks/{task_id}:
    put:
      tags: [tasks]
      operationId: updateTask
      description: Обновить задачу
      parameters:
        - $ref: "#/components/parameters/TaskId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string }
                description: { type: string }
                status: { $ref: "#/components/schemas/TaskStatus" }
                assignee_id:
                  type: integer
                  format: int64
                  nullable: true
                version: { type: integer, format: int64 }
      responses:
        "200":
          description: Задача обновлена
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Task" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: Задачу уже изменили
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    TaskId:
      name: task_id
      in: path
      required: true
      schema: { type: integer, format: int64 }
  responses:
    ValidationError:
      description: Невалидный запрос
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Неавторизован
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: Доступ запрещен
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Ресурс не найден
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
...
    TaskStatus:
      type: string
      enum: [todo, in_progress, done, cancelled]
      default: todo
    Task:
      type: object
      properties:
        id: { type: integer, format: int64 }
        team_id: { type: integer, format: int64 }
        title: { type: string }
        description: { type: string }
        status: { $ref: "#/components/schemas/TaskStatus" }
        created_by: { type: integer, format: int64 }
        assignee_id: { type: integer, format: int64, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        closed_at: { type: string, format: date-time, nullable: true }
        version: { type: integer, format: int64 }
    TaskPage:
      type: object
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Task" }
        limit: { type: integer }
        offset: { type: integer }
...